Changes to SoGtkDevice-derived classes eventhandling and signalhandler installation.  All by Ralf Corsepius.
* Instead of installing an eventhandler for all signals, install separate
eventhandlers for each signal to be monitored.  This avoids duplicate
calls to eventhandlers (A behaviour which currently happens, and probably
also is present in SGI's libSoXt).

* Evaluate SoGtkKeyboard::events to dynamically setup event-masks and
signal handlers of the calling widget.

* Install an enter-notify-eventhandler to setup CAN_FOCUS on the widget
the SoGtkKeyboard-handler gets attached to.  This implements a sort of
auto-keyboardfocusing behavior.  SoKeyboardEvents get propagated to the
eventhandler passed to SoGtkKeyboard::enable.

Two known issues:
* SoGtkKeyboard::disable deinstalls more handlers than it is supposed to.
* Missing parameter checks.